Change logback versions such that logging works again
In commit 601f2c91, the logback version was manually upgraded to address security
vulnerabilities. Unfortunately, the chosen logback version is not compatible
with the Spring Boot version used in this project, resulting in no logs being
output.

This commit adjusts the logback version override. The drawback is that
vulnerable logback versions are used again (although the versions in use no
longer contain any vulnerabilities rated as ‘High’). The clear advantage is that
logging works again.